pyproject.toml
[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[project]
name = "mjlab"
version = "0.1.0"
license = "Apache-2.0"
license-files = ["LICENSE"]
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    {name = "The MjLab Developers"},
]
keywords = ["mujoco", "mujoco-warp", "simulation", "reinforcement-learning", "robotics", "motion-imitation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Typing :: Typed",
    "Environment :: GPU :: NVIDIA CUDA",
    "Topic :: Scientific/Engineering",
    "Natural Language :: English",
]
description = "Motion Imitation framework powered by MuJoCo-Warp for RL and robotics research."
requires-python = ">=3.10,<3.14"
dependencies = [
    # 核心依赖
    "prettytable",
    "pyyaml",
    "tqdm",
    "tyro>=0.10.0a6",
    "torch>=2.7.0",
    "warp-lang>=1.10.1",
    "mujoco-warp @ git+https://github.com/google-deepmind/mujoco_warp@9fc294d86955a303619a254cefae809a41adb274",
    "mujoco<=3.3.8",
    "trimesh>=4.8.3",
    # 可视化
    "viser>=1.0.16",
    "moviepy",
    # RL训练
    "tensordict",
    "rsl-rl-lib>=3.2.0",
    "tensorboard>=2.20.0",
]

[project.urls]
"Bug Reports" = "https://github.com/mujocolab/mjlab/issues"
"Source" = "https://github.com/mujocolab/mjlab"

[project.scripts]
# 核心脚本
play = "mjlab.scripts.play:main"
sample-train = "mjlab.scripts.sample_train:main"
# 精简脚本
train = "mjlab.scripts.train_minimal:main"
sim = "mjlab.scripts.sim_minimal:main"

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "ruff>=0.14.7",
]

[project.optional-dependencies]
cu128 = ["torch>=2.7.0"]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com/"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "mujoco"
url = "https://py.mujoco.org"
explicit = true

[tool.uv.sources]
warp-lang = { index = "nvidia", marker = "sys_platform != 'darwin'" }
mujoco = { index = "mujoco" }
torch = { index = "pytorch-cu128", extra = "cu128", marker = "sys_platform != 'darwin'" }

[tool.ruff]
src = ["src"]
indent-width = 2
exclude = [
    "src/mjlab/utils/lab_api",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "B"]
ignore = ["B011"]

[tool.pyright]
pythonVersion = "3.10"
ignore = ["./src/mjlab/utils/lab_api"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = "--strict-markers"
